pyproject.toml
[tool.poetry]
name = "poketwo"
version = "0.1.0"
description = ""
authors = ["Oliver Ni <oliver.ni@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
orjson = "^3.4.6"
"discord.py-stubs" = "^1.5.1"
aio-pika = "^6.7.1"
aioredis = "^1.3.1"
jishaku = "^1.20.0"
asyncpg = "^0.21.0"
tusker = "^0.3.2"
yoyo-migrations = "^7.2.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
